refactor: merged DeleteMenu component to LayerTemplate
